backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic[email]==2.10.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.27.2
python-multipart==0.0.17
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
haversine==2.8.0
